refactor: result, symptom 컴포넌트 분리
