--- src/components/resultPage/index.tsx
@@ -1,15 +1,0 @@
-import BottomBar from "./common/BottomBar";
-import CoverPage from "./coverPage/CoverPage";
-import DefinitionPage from "./definitionPage/DefinitionPage";
-import LifePage from "./lifePage/LifePage";
-import MedicinePage from "./medicinePage/MedicinePage";
-import TreatmentPage from "./treatmentPage/TreatmentPage";
-
-export {
-  BottomBar,
-  CoverPage,
-  DefinitionPage,
-  LifePage,
-  MedicinePage,
-  TreatmentPage,
-};
