--- src/components/resultPage/lifePage/LifeComponent.tsx
@@ -1,44 +1,0 @@
-import styled from "styled-components";
-import { ILifeComponent } from "../../../interfaces/resultPage";
-import Description from "../common/Description";
-
-const Background = styled.section<{ idx: number }>`
-  display: flex;
-  background-color: ${({ theme, idx }) =>
-    idx % 2 === 0 ? theme.color.grey_800 : "transparent"};
-`;
-const Icon = styled.section`
-  margin-top: 2.1rem;
-  margin-left: 2.4rem;
-
-  font-size: 2.3rem;
-`;
-const Contents = styled.section`
-  padding: 1.6rem 2.4rem 1.6rem 1.6rem;
-  width: calc(var(--vw, 1vw) * 100 - 4.8rem);
-`;
-const Title = styled.section`
-  line-height: 150%;
-  font-size: 1.6rem;
-  font-weight: 300;
-
-  color: ${({ theme }) => theme.color.grey_200};
-
-  margin-bottom: 0.2rem;
-`;
-
-const LifeComponent = ({ idx, icon, title, content }: ILifeComponent) => {
-  return (
-    <Background idx={idx}>
-      <Icon>
-        <img alt="icon" src={icon} width={23} height={23} />
-      </Icon>
-      <Contents>
-        <Title>{title}</Title>
-        <Description text={content} />
-      </Contents>
-    </Background>
-  );
-};
-
-export default LifeComponent;
