+++ src/pages/result/description/index.tsx
@@ -1,0 +1,7 @@
+import { Text } from "./index.style";
+
+const Description = ({ text }: { text: string }) => {
+  return <Text>{text}</Text>;
+};
+
+export default Description;
