+++ src/pages/result/lifePage/index.style.tsx
@@ -1,0 +1,12 @@
+import styled from "styled-components";
+
+export const Container = styled.section`
+  padding-top: 5.6rem;
+  padding-bottom: 13rem;
+
+  .title-box {
+    margin: 2rem 2.4rem;
+
+    width: calc(var(--vw, 1vw) * 100 - 4.8rem);
+  }
+`;
