+++ src/pages/result/treatmentPage/index.style.tsx
@@ -1,0 +1,11 @@
+import styled from "styled-components";
+
+export const Container = styled.section`
+  padding-top: 5.6rem;
+  padding-bottom: 13rem;
+  margin: 0 2.4rem;
+
+  .title-box {
+    margin: 2rem 0rem;
+  }
+`;
