+++ src/pages/symptom/character/PointShader.ts
